- fixing "resource on target" with protected frames
Switching over to manual handling of the "nameplateResourceOnTarget" setting. This prevents the frame from being "forbidden" if friendly plates are shown in dungeons/instances/raids and targeted. As a result, the resources won't be shown on protected plates! The new setting is stored per character and replaces the CVar setting.