Update questInfo
Added a "progressQuest" bool to the questinfo that gets exposed to unitFrame. This will help with identifying which quests are %-based by simply checking the bool value. 

It also helps with setting the total for %-based quests by just checking the bool and doing `if progressQuest then questData.total = 100 else questData.total = amount2 end`