Fixed hot reload for the script tab
There was a bug where the scriptInfo was carrying  globalScriptObject and running scripts from this object, but it never get updated when scripts are recompiled making them always use the first compiled version when the addon load.

With this fix, constructor also needed a rework to run properly, so now globalScriptObject store the time of its latest update and ScriptGetInfo() can compare if there's a new version of scripts to run the contructor after each compilation.