Added script option to select an audio to play.
The function call to use is: Plater.PlayAudioForScript(bCanUseScriptAudio, audioFilePath, envTable)
Hook a toggle option in the bCanUseScriptAudio, the audio file option into the audioFilePath and the envTable for the script.
If the user has an audio set for the spellId in the Cast Colors tab, it'll use that, if there's no audio set, it uses the audio from script.